Rewording added for toolbar & support toolbar popup menu
