+++ reword/src/main/java/dev/b3nedikt/reword/transformer/SupportToolbarViewTransformer.kt
@@ -1,6 +1,14 @@
 package dev.b3nedikt.reword.transformer
 
+import android.content.res.Resources
+import android.util.AttributeSet
+import android.util.Pair
+import android.util.Xml
 import androidx.appcompat.widget.Toolbar
+import org.xmlpull.v1.XmlPullParser
+import org.xmlpull.v1.XmlPullParserException
+import java.io.IOException
+import java.util.HashMap
 
 /**
  * A [ViewTransformer] which transforms the androidX [Toolbar]
@@ -11,19 +19,39 @@ internal object SupportToolbarViewTransformer : AbstractViewTransformer<Toolbar>
     private const val ATTRIBUTE_SUBTITLE = "subtitle"
     private const val ATTRIBUTE_APP_TITLE = "app:title"
     private const val ATTRIBUTE_APP_SUBTITLE = "app:subtitle"
+    private const val ATTRIBUTE_ANDROID_TITLE = "android:title"
+    private const val ATTRIBUTE_MENU = "menu"
+    private const val ATTRIBUTE_APP_MENU = "app:menu"
+    private const val ATTRIBUTE_ID = "id"
+    private const val ATTRIBUTE_ANDROID_ID = "android:id"
+    private const val ATTRIBUTE_TITLE_CONDENSED = "titleCondensed"
+    private const val ATTRIBUTE_ANDROID_TITLE_CONDENSED = "android:titleCondensed"
+    private const val XML_MENU = "menu"
+    private const val XML_ITEM = "item"
 
     override val viewType = Toolbar::class.java
 
-    override val supportedAttributes = setOf(ATTRIBUTE_TITLE, ATTRIBUTE_SUBTITLE,
+    override val supportedAttributes = setOf(
+            ATTRIBUTE_MENU, ATTRIBUTE_APP_MENU, ATTRIBUTE_ID,
+            ATTRIBUTE_ANDROID_ID, ATTRIBUTE_ANDROID_TITLE, ATTRIBUTE_TITLE_CONDENSED,
+            ATTRIBUTE_ANDROID_TITLE_CONDENSED, XML_MENU, XML_ITEM, ATTRIBUTE_TITLE, ATTRIBUTE_SUBTITLE,
             ATTRIBUTE_APP_TITLE, ATTRIBUTE_APP_SUBTITLE)
 
     override fun Toolbar.transform(attrs: Map<String, Int>) {
         attrs.forEach { attribute ->
             when (attribute.key) {
-                ATTRIBUTE_TITLE, ATTRIBUTE_APP_TITLE ->
-                    updateTexts(attribute.value, this::setTitle)
-                ATTRIBUTE_SUBTITLE, ATTRIBUTE_APP_SUBTITLE ->
-                    updateTexts(attribute.value, this::setSubtitle)
+                ATTRIBUTE_TITLE, ATTRIBUTE_APP_TITLE -> updateTexts(attribute.value, this::setTitle)
+                ATTRIBUTE_SUBTITLE, ATTRIBUTE_APP_SUBTITLE -> updateTexts(attribute.value, this::setSubtitle)
+                ATTRIBUTE_MENU, ATTRIBUTE_APP_MENU -> {
+                    PopupMenuHelper.getMenuItemsStrings(resources, attrs[attribute.key] ?: 0).forEach {
+                        if (it.value.title != 0) {
+                            menu.findItem(it.key).title = resources.getString(it.value.title)
+                        }
+                        if (it.value.titleCondensed != 0) {
+                            menu.findItem(it.key).titleCondensed = resources.getString(it.value.titleCondensed)
+                        }
+                    }
+                }
             }
         }
     }
